refactor: reuse toolbar types and extract mantine context
